# docs/swagger.yaml
basePath: /
definitions:
...
  github_com_proarash_ecommerce-server_internal_media.Media:
    properties:
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      file_name:
        type: string
      file_size:
        type: integer
      id:
        type: integer
      media_type:
        type: string
      mime_type:
        type: string
      updatedAt:
        type: string
      url:
        type: string
    type: object
  github_com_proarash_ecommerce-server_internal_product.Attribute:
    properties:
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      id:
        type: integer
      key:
        type: string
      name:
        type: string
      title:
        type: string
      updatedAt:
        type: string
    type: object
  github_com_proarash_ecommerce-server_internal_product.Category:
    properties:
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      id:
        type: integer
      media:
        $ref: '#/definitions/github_com_proarash_ecommerce-server_internal_media.Media'
      media_id:
        type: integer
      name:
        type: string
      parent_id:
        type: integer
      slug:
        type: string
      updatedAt:
        type: string
    type: object
  github_com_proarash_ecommerce-server_internal_product.Product:
    properties:
      attributes:
        items:
          $ref: '#/definitions/github_com_proarash_ecommerce-server_internal_product.Attribute'
        type: array
      category:
        $ref: '#/definitions/github_com_proarash_ecommerce-server_internal_product.Category'
      category_id:
        type: integer
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      description:
        type: string
      id:
        type: integer
      is_active:
        type: boolean
      media:
        items:
          $ref: '#/definitions/github_com_proarash_ecommerce-server_internal_media.Media'
        type: array
      price:
        type: number
      sku:
        type: string
      title:
        type: string
      updatedAt:
        type: string
    type: object
...
  github_com_proarash_ecommerce-server_internal_types.ApiResponse:
    properties:
      data: {}
      message:
        type: string
    type: object
  github_com_proarash_ecommerce-server_internal_types.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  github_com_proarash_ecommerce-server_internal_types.PaginatedResponse:
    properties:
      items: {}
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
    type: object
...
  gorm.DeletedAt:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
...
  internal_finance.CreatePreInvoiceDto:
    properties:
      amount:
        type: number
      order_id:
        minimum: 1
        type: integer
    required:
    - order_id
    type: object
  internal_finance.Order:
    properties:
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      discount_amount:
        type: number
      discount_id:
        type: integer
      id:
        type: integer
      items:
        items:
          $ref: '#/definitions/internal_finance.OrderItem'
        type: array
      status:
        enum:
        - pending
        - paid
        - failed
        - cancelled
        - processing
        - delivered
        type: string
      total_amount:
        type: number
      updatedAt:
        type: string
      user_id:
        type: integer
    type: object
  internal_finance.OrderItem:
    properties:
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      id:
        type: integer
      order_id:
        type: integer
      product:
        $ref: '#/definitions/github_com_proarash_ecommerce-server_internal_product.Product'
      product_id:
        type: integer
      quantity:
        type: integer
      unit_price:
        type: number
      updatedAt:
        type: string
    type: object
  internal_finance.PreInvoice:
    properties:
      amount:
        type: number
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      id:
        type: integer
      invoice_number:
        type: string
      issued_by:
        type: string
      order_id:
        type: integer
      status:
        enum:
        - issued
        - paid
        - cancelled
        type: string
      updatedAt:
        type: string
      user_id:
        type: integer
    type: object
  internal_finance.ReportResponse:
    properties:
      orders_count:
        type: integer
      period:
        enum:
        - daily
        - weekly
        - monthly
        - annually
        type: string
      rows:
        items:
          $ref: '#/definitions/internal_finance.ReportRow'
        type: array
      total_amount:
        type: number
    type: object
  internal_finance.ReportRow:
    properties:
      orders_count:
        type: integer
      period_start:
        type: string
      total_amount:
        type: number
    type: object
...
host: localhost:4000
info:
  contact: {}
  description: 'E-commerce backend: catalog, cart, orders, Zibal payments, inventory,
    CMS, support chat and notifications. All JSON responses are wrapped in {"message":
    string, "data": any}.'
  title: E-Commerce Server API
  version: "1.0"
paths:
...
  /finance/preinvoices:
    get:
      parameters:
      - description: Session to authenticate with; selects the staff or customer auth
          cookie
        enum:
        - staff
        - customer
        in: query
        name: user_type
        type: string
      - description: Status
        enum:
        - issued
        - paid
        - cancelled
        in: query
        name: status
        type: string
      - description: User ID
        in: query
        name: user_id
        type: integer
      - default: 1
        description: Page
        in: query
        name: page
        type: integer
      - default: 20
        description: Limit
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_proarash_ecommerce-server_internal_types.ApiResponse'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/github_com_proarash_ecommerce-server_internal_types.PaginatedResponse'
                  - properties:
                      items:
                        items:
                          $ref: '#/definitions/internal_finance.PreInvoice'
                        type: array
                    type: object
              type: object
        "403":
          description: Forbidden
          schema:
            allOf:
            - $ref: '#/definitions/github_com_proarash_ecommerce-server_internal_types.ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_proarash_ecommerce-server_internal_types.ErrorResponse'
              type: object
      security:
      - BearerAuth: []
      summary: List pre-invoices
      tags:
      - Finance
    post:
      consumes:
      - application/json
      description: Issues a pre-invoice for an order (amount defaults to the order
        total) and notifies the customer
      parameters:
      - description: Session to authenticate with; selects the staff or customer auth
          cookie
        enum:
        - staff
        - customer
        in: query
        name: user_type
        type: string
      - description: Pre-invoice
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_finance.CreatePreInvoiceDto'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/github_com_proarash_ecommerce-server_internal_types.ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/internal_finance.PreInvoice'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/github_com_proarash_ecommerce-server_internal_types.ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_proarash_ecommerce-server_internal_types.ErrorResponse'
              type: object
        "404":
          description: Not Found
          schema:
            allOf:
            - $ref: '#/definitions/github_com_proarash_ecommerce-server_internal_types.ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_proarash_ecommerce-server_internal_types.ErrorResponse'
              type: object
      security:
      - BearerAuth: []
      summary: Create manual pre-invoice
      tags:
      - Finance
  /finance/reports:
    get:
      description: Paid, processing and delivered orders aggregated per period bucket
      parameters:
      - description: Session to authenticate with; selects the staff or customer auth
          cookie
        enum:
        - staff
        - customer
        in: query
        name: user_type
        type: string
      - default: daily
        description: Bucket size
        enum:
        - daily
        - weekly
        - monthly
        - annually
        in: query
        name: period
        type: string
      - description: From date (YYYY-MM-DD)
        in: query
        name: from
        type: string
      - description: To date inclusive (YYYY-MM-DD)
        in: query
        name: to
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_proarash_ecommerce-server_internal_types.ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/internal_finance.ReportResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/github_com_proarash_ecommerce-server_internal_types.ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_proarash_ecommerce-server_internal_types.ErrorResponse'
              type: object
        "403":
          description: Forbidden
          schema:
            allOf:
            - $ref: '#/definitions/github_com_proarash_ecommerce-server_internal_types.ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_proarash_ecommerce-server_internal_types.ErrorResponse'
              type: object
      security:
      - BearerAuth: []
      summary: Sales report
      tags:
      - Finance
...
  /support/orders:
    get:
      parameters:
      - description: Session to authenticate with; selects the staff or customer auth
          cookie
        enum:
        - staff
        - customer
        in: query
        name: user_type
        type: string
      - description: Status
        in: query
        name: status
        type: string
      - description: User ID
        in: query
        name: user_id
        type: integer
      - default: 1
        description: Page
        in: query
        name: page
        type: integer
      - default: 20
        description: Limit
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_proarash_ecommerce-server_internal_types.ApiResponse'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/github_com_proarash_ecommerce-server_internal_types.PaginatedResponse'
                  - properties:
                      items:
                        items:
                          $ref: '#/definitions/internal_finance.Order'
                        type: array
                    type: object
              type: object
        "403":
          description: Forbidden
          schema:
            allOf:
            - $ref: '#/definitions/github_com_proarash_ecommerce-server_internal_types.ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_proarash_ecommerce-server_internal_types.ErrorResponse'
              type: object
      security:
      - BearerAuth: []
      summary: Read-only customer orders
      tags:
      - Support
  /support/preinvoices:
    get:
      parameters:
      - description: Session to authenticate with; selects the staff or customer auth
          cookie
        enum:
        - staff
        - customer
        in: query
        name: user_type
        type: string
      - description: Status
        in: query
        name: status
        type: string
      - description: User ID
        in: query
        name: user_id
        type: integer
      - default: 1
        description: Page
        in: query
        name: page
        type: integer
      - default: 20
        description: Limit
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_proarash_ecommerce-server_internal_types.ApiResponse'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/github_com_proarash_ecommerce-server_internal_types.PaginatedResponse'
                  - properties:
                      items:
                        items:
                          $ref: '#/definitions/internal_finance.PreInvoice'
                        type: array
                    type: object
              type: object
        "403":
          description: Forbidden
          schema:
            allOf:
            - $ref: '#/definitions/github_com_proarash_ecommerce-server_internal_types.ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_proarash_ecommerce-server_internal_types.ErrorResponse'
              type: object
      security:
      - BearerAuth: []
      summary: Read-only customer pre-invoices
      tags:
      - Support
...
  /user/orders:
    get:
      parameters:
      - description: Session to authenticate with; selects the staff or customer auth
          cookie
        enum:
        - staff
        - customer
        in: query
        name: user_type
        type: string
      - description: Status
        in: query
        name: status
        type: string
      - default: 1
        description: Page
        in: query
        name: page
        type: integer
      - default: 20
        description: Limit
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_proarash_ecommerce-server_internal_types.ApiResponse'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/github_com_proarash_ecommerce-server_internal_types.PaginatedResponse'
                  - properties:
                      items:
                        items:
                          $ref: '#/definitions/internal_finance.Order'
                        type: array
                    type: object
              type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
            - $ref: '#/definitions/github_com_proarash_ecommerce-server_internal_types.ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_proarash_ecommerce-server_internal_types.ErrorResponse'
              type: object
      security:
      - BearerAuth: []
      summary: Customer orders
      tags:
      - User
  /user/orders/{id}:
    get:
      parameters:
      - description: Session to authenticate with; selects the staff or customer auth
          cookie
        enum:
        - staff
        - customer
        in: query
        name: user_type
        type: string
      - description: Order ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_proarash_ecommerce-server_internal_types.ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/internal_finance.Order'
              type: object
        "404":
          description: Not Found
          schema:
            allOf:
            - $ref: '#/definitions/github_com_proarash_ecommerce-server_internal_types.ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_proarash_ecommerce-server_internal_types.ErrorResponse'
              type: object
      security:
      - BearerAuth: []
      summary: Customer order details
      tags:
      - User
  /user/preinvoices:
    get:
      parameters:
      - description: Session to authenticate with; selects the staff or customer auth
          cookie
        enum:
        - staff
        - customer
        in: query
        name: user_type
        type: string
      - default: 1
        description: Page
        in: query
        name: page
        type: integer
      - default: 20
        description: Limit
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_proarash_ecommerce-server_internal_types.ApiResponse'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/github_com_proarash_ecommerce-server_internal_types.PaginatedResponse'
                  - properties:
                      items:
                        items:
                          $ref: '#/definitions/internal_finance.PreInvoice'
                        type: array
                    type: object
              type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
            - $ref: '#/definitions/github_com_proarash_ecommerce-server_internal_types.ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_proarash_ecommerce-server_internal_types.ErrorResponse'
              type: object
      security:
      - BearerAuth: []
      summary: Customer pre-invoices
      tags:
      - User
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and the JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"